requirements.txt
faker
google-cloud